requirements.txt
Django==5.2.8
djangorestframework==3.16.1
django-cors-headers==4.9.0
django-filter==24.3
Pillow==12.0.0
python-dotenv==1.0.0
pytest==8.3.4
pytest-django==4.9.0
pytest-cov==6.0.0
factory-boy==3.3.1
drf-spectacular==0.29.0

